Changed GetIdsByName to generic
