+++ StockManagement/Functions/CalculateStock.cs
