+++ StockManagement/Functions/Search.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace StockManagement
+{
+    public class Search
+    {
+        public List<int?> GetIdsByName<T>(List<T> stock, string? name) where T: Stock
+        {
+            if (String.IsNullOrEmpty(name) == false)
+            {
+                List<int?> items = stock.Where(x => x.Name == name).Select(x => x.Id).ToList();
+                return items;
+            }
+            return null;
+        }
+    }
+}
