+++ StockManagement/Repositories/IRepository.cs
