+++ StockManagement/Repositories/LaptopRepository.cs
