+++ StockManagement/Stock/Laptop.cs
